updated API signup/login route
