+++ app/api/auth/sendVerificationCode/route.ts
@@ -1,12 +1,18 @@
 import { generateVerificationCode } from "@/helpers/common";
-import { sendEmail } from "@/helpers/emails";
+import { sendEmail, sendVerificationCode } from "@/helpers/emails";
 import { prisma } from "@/prisma/prisma";
+import { isUserAlreadyExists } from "@/services/membership/signup";
 import moment from "moment";
 import { NextRequest } from "next/server";
 
 export const POST = async (req: NextRequest) => {
   try {
     const { email } = await req.json();
+
+    const isValidEmail = await isUserAlreadyExists(email, false);
+    if (!isValidEmail) {
+      return Response.json({ error: "Account doesn't exsists!" }, { status: 404 });
+    }
 
     const verificationCode = generateVerificationCode();
 
@@ -19,7 +25,7 @@ export const POST = async (req: NextRequest) => {
     });
 
     // Send verification code to user
-    const messageId = await sendEmail({ email, subject: "Login - Verification Code", body: `<p>${verificationCode}</p>` });
+    const messageId = await sendVerificationCode(email);
 
     return Response.json({ message: "Verification code sent!", messageId }, { status: 200 });
   } catch (err: any) {
