+++ helpers/emails.tsx
@@ -1,4 +1,7 @@
 import { ServerClient } from "postmark";
+import { generateVerificationCode } from "./common";
+import moment from "moment";
+import { prisma } from "@/prisma/prisma";
 
 const client = new ServerClient(process.env.POSTMARK_SERVER_TOKEN!);
 
@@ -10,4 +13,20 @@ export const sendEmail = async ({ email, subject, body }: any) => {
     HtmlBody: body,
   });
   return res.MessageID;
+};
+
+export const sendVerificationCode = async (email: string) => {
+  const verificationCode = generateVerificationCode();
+
+  const expiryTime = moment().add(5, "minutes").toDate(); // expiry time for verification code
+
+  await prisma.verificationToken.upsert({
+    where: { email },
+    create: { email, token: verificationCode, expires: expiryTime },
+    update: { token: verificationCode, expires: expiryTime },
+  });
+
+  // Send verification code to user
+  const messageId = await sendEmail({ email, subject: "Login - Verification Code", body: `<p>${verificationCode}</p>` });
+  return messageId;
 };
